refactor code, add postCont, postDisc for simulation
