+++ demo/tacas2023/exp1/exp1_sim.py
@@ -59,7 +59,7 @@ if __name__ == "__main__":
     # traces.dump("./output1.json")
     # traces = AnalysisTree.load('./output1.json')
     start_time = time.time()
-    traces = scenario.simulate(40, time_step, seed=4)
+    traces = scenario.simulate_simple(40, time_step, seed=4)
     run_time = time.time() - start_time
     traces.dump('demo/tacas2023/exp1/output1_sim.json')
     # print({
