+++ verse/analysis/simulator.py
@@ -1,11 +1,17 @@
-from typing import List, Tuple
+from typing import List, Tuple, Dict, DefaultDict, Any, NamedTuple
 import copy
 import itertools
 import functools
+from collections import defaultdict, namedtuple
+import warnings
+import types
+import sys
 
+from enum import Enum
+from verse.analysis.incremental import CachedRTTrans, CachedSegment, combine_all, reach_trans_suit, sim_trans_suit
 import pprint
 from verse.agents.base_agent import BaseAgent
-
+from verse.analysis.utils import EGO,OTHERS
 from verse.analysis.incremental import SimTraceCache, convert_sim_trans, to_simulate
 from verse.analysis.utils import dedup
 from verse.parser.parser import ModePath, find
@@ -16,9 +22,140 @@ from verse.analysis.analysis_tree import AnalysisTreeNode, AnalysisTree
 
 PathDiffs = List[Tuple[BaseAgent, ModePath]]
 
-def red(s):
-    return "\x1b[31m" + s + "\x1b[0m" #]]
+def namedTupleToClass(inp: NamedTuple):
+    res = types.SimpleNamespace()
+    for field in inp._fields:
+        setattr(res, field, getattr(inp, field))
+    return res
 
+def convertStrToEnum(inp, agent:BaseAgent, dl):
+    res = inp
+    for field in res.__dict__:
+        for state_def_name in agent.decision_logic.state_defs:
+            if field in agent.decision_logic.state_defs[state_def_name].disc:
+                idx = agent.decision_logic.state_defs[state_def_name].disc.index(field)
+                field_type = agent.decision_logic.state_defs[state_def_name].disc_type[idx]
+                enum_class = getattr(dl, field_type)
+                setattr(res, field, enum_class[getattr(res, field)])
+    return res
+
+def convertEnumToStr(inp, agent:BaseAgent, dl):
+    res = inp
+    for field in res.__dict__:
+        val = getattr(res, field)
+        if isinstance(val, Enum):
+            setattr(res, field, val.name)
+    return res
+
+def is_equivalent(inp1:NamedTuple, inp2:NamedTuple):
+    for field in inp1._fields:
+        if getattr(inp1, field) != getattr(inp2, field):
+            return False
+    return True
+
+def disc_field(field:str, agent:BaseAgent):
+    for state_def_name in agent.decision_logic.state_defs:
+        state_def = agent.decision_logic.state_defs[state_def_name]
+        if field in state_def.disc:
+            return True 
+    return False
+    
+def pack_env(agent: BaseAgent, ego_ty_name: str, cont: Dict[str, float], disc: Dict[str, str], track_map):
+    state_ty = None #namedtuple(ego_ty_name, agent.decision_logic.state_defs[ego_ty_name].all_vars())
+    packed: DefaultDict[str, Any] = defaultdict(dict)
+    # packed = {}
+    for e in [cont, disc]:
+        for k, v in e.items():
+            k1, k2 = k.split(".")
+            packed[k1][k2] = v
+    ego_keys, ego_vals = tuple(map(list, zip(*packed[EGO].items())))
+    state_ty = namedtuple(ego_ty_name, ego_keys)
+    for arg in agent.decision_logic.args:
+        if "map" in arg.name:
+            packed[arg.name] = track_map
+        elif arg.name != EGO:
+            other = arg.name
+            if other in packed:
+                other_keys, other_vals = tuple(map(list, zip(*packed[other].items())))
+                state_ty = namedtuple(ego_ty_name, other_keys)
+                packed[other] = list(map(lambda v: state_ty(*v), zip(*other_vals)))
+                if not arg.is_list:
+                    packed[other] = packed[other][0]
+            else:
+                if arg.is_list:
+                    packed[other] = []
+                else:
+                    raise ValueError(f"Expected one {ego_ty_name} for {other}, got none")
+
+    packed[EGO] = state_ty(**packed[EGO])
+    return dict(packed.items())
+
+def check_sim_transitions(agent: BaseAgent, guards: List[Tuple], cont, disc, map, state, mode):
+    asserts = []
+    satisfied_guard = []
+    agent_id = agent.id
+    # Unsafety checking
+    ego_ty_name = find(agent.decision_logic.args, lambda a: a.name == EGO).typ
+    packed_env = pack_env(agent, ego_ty_name, cont, disc, map)
+
+    # Check safety conditions
+    for assertion in agent.decision_logic.asserts:
+        if eval(assertion.pre, packed_env):
+            if not eval(assertion.cond, packed_env):
+                del packed_env["__builtins__"]
+                print(f"assert hit for {agent_id}: \"{assertion.label}\" @ {packed_env}")
+                asserts.append(assertion.label)
+    if len(asserts) != 0:
+        return asserts, satisfied_guard
+
+    all_resets = defaultdict(list)
+    for path, disc_vars in guards:
+        env = pack_env(agent, ego_ty_name, cont, disc, map)    # TODO: diff disc -> disc_vars?
+
+        # Collect all the hit guards for this agent at this time step
+        if eval(path.cond, env):
+            # If the guard can be satisfied, handle resets
+            all_resets[path.var].append((path.val, path))
+
+    iter_list = []
+    for vals in all_resets.values():
+        paths = [p for _, p in vals]
+        iter_list.append(zip(range(len(vals)), paths))
+    pos_list = list(itertools.product(*iter_list))
+    if len(pos_list) == 1 and pos_list[0] == ():
+        return None, satisfied_guard
+    for pos in pos_list:
+        next_init = copy.deepcopy(state)
+        dest = copy.deepcopy(mode)
+        possible_dest = [[elem] for elem in dest]
+        for j, (reset_idx, path) in enumerate(pos):
+            reset_variable = list(all_resets.keys())[j]
+            res = eval(all_resets[reset_variable][reset_idx][0], packed_env)
+            ego_type = agent.decision_logic.state_defs[ego_ty_name]
+            if "mode" in reset_variable:
+                var_loc = ego_type.disc.index(reset_variable)
+                assert not isinstance(res, list), res
+                possible_dest[var_loc] = [(res, path)]
+            else:
+                var_loc = ego_type.cont.index(reset_variable)
+                next_init[var_loc] = res
+        all_dest = list(itertools.product(*possible_dest))
+        if not all_dest:
+            warnings.warn(
+                f"Guard hit for mode {mode} for agent {agent_id} without available next mode")
+            all_dest.append(None)
+        for dest in all_dest:
+            assert isinstance(dest, tuple)
+            paths = []
+            pure_dest = []
+            for d in dest:
+                if isinstance(d, tuple):
+                    pure_dest.append(d[0])
+                    paths.append(d[1])
+                else:
+                    pure_dest.append(d)
+            satisfied_guard.append((agent_id, pure_dest, next_init, paths))
+    return None, satisfied_guard
 
 class Simulator:
     def __init__(self, config):
@@ -26,6 +163,220 @@ class Simulator:
         self.cache = SimTraceCache()
         self.config = config
         self.cache_hits = (0, 0)
+
+    def postCont(
+        self, 
+        node: AnalysisTreeNode,
+        remain_time: float,
+        time_step: float,
+        track_map
+    ) -> AnalysisTreeNode:
+        for agent_id in node.agent:
+            mode = node.mode[agent_id]
+            init = node.init[agent_id]
+            # pp(("sim", agent_id, *mode, *init))
+            # Simulate the trace starting from initial condition
+            trace = node.agent[agent_id].TC_simulate(
+                mode, init, remain_time, time_step, track_map)
+            trace[:, 0] += node.start_time
+            trace = trace.tolist()
+            node.trace[agent_id] = trace
+        return node
+
+    def postDisc(
+        self,
+        node: AnalysisTreeNode,
+        track_map, 
+        sensor,
+        cache: Dict = {},
+        paths: List = []
+    ):
+
+        trace_length = len(list(node.trace.values())[0])
+
+        # For each agent
+        agent_guard_dict = defaultdict(list)
+        cached_guards = defaultdict(list)
+        min_trans_ind = None
+        cached_trans = defaultdict(list)
+
+        if not cache:
+            paths = [(agent, p) for agent in node.agent.values() for p in agent.decision_logic.paths]
+        else:
+            _transitions = [(aid, trans) for aid, seg in cache.items() for trans in seg.transitions if sim_trans_suit(trans.inits, node.init)]
+            # pp(("cached trans", _transitions))
+            if len(_transitions) > 0:
+                min_trans_ind = min([t.transition for _, t in _transitions])
+                # pp(("min", min_trans_ind))
+                for aid, trans in _transitions:
+                    # TODO: check for asserts
+                    if trans.transition == min_trans_ind:
+                        # pp(("chosen tran", aid, trans))
+                        cached_trans[aid].append((aid, trans.disc, trans.cont, trans.paths))
+                for agent_id in cached_trans:
+                    cached_trans[agent_id] = dedup(cached_trans[agent_id], lambda p: p[:3])
+                if len(paths) == 0:
+                    # print(red("full cache"))
+                    return None, dict(cached_trans), min_trans_ind
+
+                path_transitions = defaultdict(int)
+                for seg in cache.values():
+                    for tran in seg.transitions:
+                        for p in tran.paths:
+                            path_transitions[p.cond] = max(path_transitions[p.cond], tran.transition)
+                for agent_id, segment in cache.items():
+                    agent = node.agent[agent_id]
+                    if len(agent.decision_logic.args) == 0:
+                        continue
+                    state_dict = {aid: (node.trace[aid][0], node.mode[aid], node.static[aid]) for aid in node.agent}
+                    agent_paths = dedup([p for tran in segment.transitions for p in tran.paths], lambda i: (i.var, i.cond, i.val))
+                    cont_var_dict_template, discrete_variable_dict, len_dict = sensor.sense(self, agent, state_dict, track_map)
+                    for path in agent_paths:
+                        cached_guards[agent_id].append((path, discrete_variable_dict, path_transitions[path.cond]))
+
+        for agent, path in paths:
+            # Get guard
+            if len(agent.decision_logic.args) == 0:
+                continue
+            agent_id = agent.id
+            agent_mode = node.mode[agent_id]
+            state_dict = {aid: (node.trace[aid][0], node.mode[aid], node.static[aid]) for aid in node.agent}
+            cont_var_dict_template, discrete_variable_dict, len_dict = sensor.sense(self, agent, state_dict, track_map)
+            agent_guard_dict[agent_id].append((path, discrete_variable_dict))
+
+        transitions = defaultdict(list)
+        # TODO: We can probably rewrite how guard hit are detected and resets are handled for simulation
+        for idx in range(trace_length):
+            if min_trans_ind != None and idx >= min_trans_ind:
+                return None, dict(cached_trans), min_trans_ind
+            satisfied_guard = []
+            all_asserts = defaultdict(list)
+            for agent_id in agent_guard_dict:
+                agent: BaseAgent = node.agent[agent_id]
+                state_dict = {aid: (node.trace[aid][idx], node.mode[aid], node.static[aid]) for aid in node.agent}
+                agent_state, agent_mode, agent_static = state_dict[agent_id]
+                agent_state = agent_state[1:]
+                continuous_variable_dict, orig_disc_vars, _ = sensor.sense(self, agent, state_dict, track_map)
+                unchecked_cache_guards = [g[:2] for g in cached_guards[agent_id] if g[2] < idx]     # FIXME: off by 1?
+                asserts, satisfied = check_sim_transitions(agent, agent_guard_dict[agent_id] + unchecked_cache_guards, continuous_variable_dict, orig_disc_vars, track_map, agent_state, agent_mode)
+                if asserts != None:
+                    all_asserts[agent_id] = asserts
+                    continue
+                if len(satisfied) != 0:
+                    satisfied_guard.extend(satisfied)
+                    # assert all(len(s[2]) == 4 for s in satisfied)
+            if len(all_asserts) > 0:
+                return all_asserts, dict(transitions), idx
+            if len(satisfied_guard) > 0:
+                for agent_idx, dest, next_init, paths in satisfied_guard:
+                    assert isinstance(paths, list)
+                    dest = tuple(dest)
+                    src_mode = node.get_mode(agent_idx, node.mode[agent_idx])
+                    src_track = node.get_track(agent_idx, node.mode[agent_idx])
+                    dest_mode = node.get_mode(agent_idx, dest)
+                    dest_track = node.get_track(agent_idx, dest)
+                    # pp(("dbg", src_track, src_mode, dest, dest_mode, dest_track))
+                    # pp((track_map.h(src_track, src_mode, dest_mode)))
+                    if not track_map or dest_track == track_map.h(src_track, src_mode, dest_mode):
+                        transitions[agent_idx].append((agent_idx, dest, next_init, paths))
+                # print("transitions", transitions)
+                break
+        return None, dict(transitions), idx
+
+    def postDisc_Simple(
+        self,
+        node: AnalysisTreeNode,
+        track_map,
+        sensor
+    ):
+        trace_length = len(list(node.trace.values())[0])
+
+        # For each agent
+        agent_guard_dict = defaultdict(list)
+        
+        paths = [(agent, p) for agent in node.agent.values() for p in agent.decision_logic.paths]
+        
+        for agent, path in paths:
+            # Get guard
+            if len(agent.decision_logic.args) == 0:
+                continue
+            agent_id = agent.id
+            state_dict = {aid: (node.trace[aid][0], node.mode[aid], node.static[aid]) for aid in node.agent}
+            cont_var_dict_template, discrete_variable_dict, len_dict = sensor.sense(self, agent, state_dict, track_map)
+            agent_guard_dict[agent_id].append((path, discrete_variable_dict))
+
+        transitions = defaultdict(list)
+        # TODO: We can probably rewrite how guard hit are detected and resets are handled for simulation
+        for idx in range(trace_length):
+            state_dict = {aid: (node.trace[aid][idx], node.mode[aid], node.static[aid]) for aid in node.agent}
+            satisfied_guard = []
+            all_asserts = defaultdict(list)
+            for agent_id in agent_guard_dict:
+                # Get agent controller 
+                # Reference: https://stackoverflow.com/questions/55905240/python-dynamically-import-modules-code-from-string-with-importlib
+                agent: BaseAgent = node.agent[agent_id]
+                dl = types.ModuleType('dl')
+                exec(agent.decision_logic.controller_code,dl.__dict__)
+                
+                # Get the input arguments for the controller function
+                # Pack the environment (create ego and others list)
+                continuous_variable_dict, orig_disc_vars, _ = sensor.sense(self, agent, state_dict, track_map)
+                arg_list = []
+                env = pack_env(agent, EGO, continuous_variable_dict, orig_disc_vars, track_map)
+                for arg in agent.decision_logic.args:
+                    if arg.name == EGO:
+                        ego = namedTupleToClass(env[EGO])
+                        ego = convertStrToEnum(ego, agent, dl)
+                        arg_list.append(ego)
+                    elif arg.name == 'track_map':
+                        arg_list.append(track_map)
+                    else:
+                        if isinstance(env[arg.name], list):
+                            tmp_list = []
+                            for item in env[arg.name]:
+                                tmp = namedTupleToClass(item)
+                                tmp = convertStrToEnum(tmp, agent, dl)
+                                tmp_list.append(tmp)
+                            arg_list.append(tmp_list)
+                        else:
+                            tmp = namedTupleToClass(env[arg.name])
+                            tmp = convertStrToEnum(tmp, agent, dl)
+                            arg_list.append(tmp)
+
+                try:
+                    # Input the environment into the actual controller
+                    output = dl.decisionLogic(*arg_list)   
+                    output = convertEnumToStr(output, agent, dl)   
+                    # Check if output is the same as ego
+                    if not is_equivalent(env[EGO], output):
+                        # If not, a transition happen, get source and destination, break
+                        next_init = []
+                        pure_dest = []
+                        for field in env[EGO]._fields:
+                            if disc_field(field, agent):
+                                pure_dest.append(getattr(output, field))
+                            else:
+                                next_init.append(getattr(output, field))  
+                        satisfied_guard.append((agent_id, pure_dest, next_init, []))
+                except AssertionError:
+                    # If assertion error happen, means assert hit happen
+                    # Get the which assert hit happen and return
+                    _, error, _ = sys.exc_info()
+                    assertion_label = error.args[0]
+                    all_asserts[agent_id] = assertion_label
+                    continue
+                
+            if len(all_asserts) > 0:
+                for agent_idx, dest, next_init, paths in satisfied_guard:
+                    transitions[agent_idx].append((agent_idx, dest, next_init, paths))
+                return all_asserts, dict(transitions), idx
+            if len(satisfied_guard)>0:
+                break
+            # Convert output to asserts, transition and idx
+        for agent_idx, dest, next_init, paths in satisfied_guard:
+            transitions[agent_idx].append((agent_idx, dest, next_init, paths))
+        return None, dict(transitions), idx
+
 
     def simulate(self, init_list, init_mode_list, static_list, uncertain_param_list, agent_list,
                  transition_graph, time_horizon, time_step, lane_map, run_num, past_runs):
@@ -107,7 +458,8 @@ class Simulator:
                 # else:
                 #     print("!!!")
 
-            asserts, transitions, transition_idx = transition_graph.get_transition_simulate(new_cache, paths_to_sim, node)
+            # asserts, transitions, transition_idx = transition_graph.get_transition_simulate(new_cache, paths_to_sim, node)
+            asserts, transitions, transition_idx = self.postDisc(node, transition_graph.map, transition_graph.sensor, new_cache, paths_to_sim)
             # pp(("transitions:", transition_idx, transitions))
 
             node.assert_hits = asserts
@@ -256,7 +608,8 @@ class Simulator:
             # TODO: for now, make sure all the segments comes from the same node; maybe we can do
             # something to combine results from different nodes in the future
             
-            asserts, transitions, transition_idx = transition_graph.get_transition_simulate_simple(node)
+            # asserts, transitions, transition_idx = transition_graph.get_transition_simulate_simple(node)
+            asserts, transitions, transition_idx = self.postDisc_Simple(node,transition_graph.map, transition_graph.sensor)
             # pp(("transitions:", transition_idx, transitions))
 
             node.assert_hits = asserts
